Add files via upload
this script extract data from a .csv file called jude_cod.csv. Extract the species whose dynamcs we are interested in modelling  and the variables that affects the abundance of species. One of the variable in Question is Temperature. We use the bottom teperature measurements. Instead of using the time series for temperature in jude_cod.csv, we use the ROMS temperature values, found in the .csv file haul_ROMS.csv.